indexdata: store fileNameNgrams in compressed posting list form.
Storing them decompressed consumes significant amounts of memory, and
the time savings is irrelevant-- it's only used for file: queries, and
decompressing posting lists while building matches is *already* used
extensively for normal search operations.

Corpus: fileNameNgrams memory usage goes from 15% of the total (3.3GB)
to 5.7% (1.11GB).

+++ read_test.go
@@ -104,7 +104,7 @@ func TestReadWriteNames(t *testing.T) {
 	if !reflect.DeepEqual([]uint32{0, 4}, data.fileNameIndex) {
 		t.Errorf("got index %v, want {0,4}", data.fileNameIndex)
 	}
-	if got := data.fileNameNgrams[stringToNGram("bCd")]; !reflect.DeepEqual(got, []uint32{1}) {
+	if got := data.fileNameNgrams[stringToNGram("bCd")]; !reflect.DeepEqual(got, []byte{1}) {
 		t.Errorf("got trigram bcd at bits %v, want sz 2", data.fileNameNgrams)
 	}
 }
